services/worker/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
tenacity==8.2.3
dbt-postgres==1.7.4
pytest==8.0.0
